Range of years implementation - NO COMMENTS
This commit begins the implementation for the range of years using 2
different spinners for Willie's Cycles app. The only thing left to
implement is the actual search for the parts within the given range of
years instead of the lower year. The 2 year spinners as well as the part
spinner have been properly implemented.